Reorganize package stats code
